rename to otelconfig
