+++ otelconfig/version.go
@@ -1,3 +1,3 @@
-package launcher
+package otelconfig
 
 const version = "1.7.0"
